+++ sample/rpcExample/FibProcessor.cs
@@ -1,6 +1,6 @@
 using System;
 using Microsoft.Extensions.Logging;
-using Softplan.Common.Messaging.RabbitMq.Abstractions;
+using Softplan.Common.Messaging.RabbitMq.Abstractions.Interfaces;
 
 namespace rpcExample
 {
