+++ sample/simplePubSub/Program.cs
@@ -2,10 +2,10 @@
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Configuration;
 using System.IO;
-using rpcExample.Properties;
+using simplePubSub.Properties;
 using Softplan.Common.Messaging;
 using Softplan.Common.Messaging.RabbitMq;
-using Softplan.Common.Messaging.RabbitMq.Abstractions;
+using Softplan.Common.Messaging.RabbitMq.Abstractions.Interfaces;
 
 namespace simplePubSub
 {
