+++ src/Softplan.Common.Messaging.RabbitMq.Abstractions/Interfaces/IConsumer.cs
@@ -1,4 +1,4 @@
-namespace Softplan.Common.Messaging.RabbitMq.Abstractions
+namespace Softplan.Common.Messaging.RabbitMq.Abstractions.Interfaces
 {
     public interface IConsumer
     {
