+++ src/Softplan.Common.Messaging.RabbitMq.Abstractions/Interfaces/IMessage.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Softplan.Common.Messaging.RabbitMq.Abstractions
+namespace Softplan.Common.Messaging.RabbitMq.Abstractions.Interfaces
 {
     public interface IMessage
     {
