+++ src/Softplan.Common.Messaging.RabbitMq.Abstractions/Interfaces/IMessagingManager.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Softplan.Common.Messaging.RabbitMq.Abstractions
+namespace Softplan.Common.Messaging.RabbitMq.Abstractions.Interfaces
 {
     public interface IMessagingManager
     {
