+++ src/Softplan.Common.Messaging.RabbitMq.Abstractions/Interfaces/IProcessor.cs
@@ -1,7 +1,7 @@
 using System;
 using Microsoft.Extensions.Logging;
 
-namespace Softplan.Common.Messaging.RabbitMq.Abstractions
+namespace Softplan.Common.Messaging.RabbitMq.Abstractions.Interfaces
 {
     public interface IProcessor
     {
