+++ src/Softplan.Common.Messaging.RabbitMq.Abstractions/Interfaces/IQueueAPIManager.cs
@@ -1,4 +1,4 @@
-namespace Softplan.Common.Messaging.RabbitMq.Abstractions
+namespace Softplan.Common.Messaging.RabbitMq.Abstractions.Interfaces
 {
     public interface IQueueApiManager
     {
