+++ src/Softplan.Common.Messaging.RabbitMq/RabbitMqPublisher.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using RabbitMQ.Client;
 using RabbitMQ.Client.Events;
-using Softplan.Common.Messaging.RabbitMq.Abstractions;
+using Softplan.Common.Messaging.RabbitMq.Abstractions.Interfaces;
 using Softplan.Common.Messaging.RabbitMq.Properties;
 
 namespace Softplan.Common.Messaging.RabbitMq
