+++ src/Softplan.Common.Messaging/Messages/ErrorMessage.cs
@@ -1,4 +1,5 @@
 using Softplan.Common.Messaging.RabbitMq.Abstractions;
+using Softplan.Common.Messaging.RabbitMq.Abstractions.Interfaces;
 
 namespace Softplan.Common.Messaging.Messages
 {
