+++ test/Softplan.Common.Messaging.RabbitMq.Tests/AMQPPublisherTest.cs
@@ -6,6 +6,7 @@ using FluentAssertions;
 using Moq;
 using RabbitMQ.Client;
 using Softplan.Common.Messaging.RabbitMq.Abstractions;
+using Softplan.Common.Messaging.RabbitMq.Abstractions.Interfaces;
 using Softplan.Common.Messaging.RabbitMq.Properties;
 using Xunit;
