+++ test/Softplan.Common.Messaging.Tests/TestProcessors/InvalidConstructorProcessor.cs
@@ -1,6 +1,6 @@
 using System;
 using Microsoft.Extensions.Logging;
-using Softplan.Common.Messaging.RabbitMq.Abstractions;
+using Softplan.Common.Messaging.RabbitMq.Abstractions.Interfaces;
 
 namespace Softplan.Common.Messaging.Tests.TestProcessors
 {
